+++ lib/jwtUtils.js
@@ -1,7 +1,7 @@
 import dotenv from 'dotenv';
 import jwt from 'jsonwebtoken';
 import consoleLogger from './consoleLogger.js';
-import { selectToken } from '../models/accessTokenTable.js';
+import { insertToken, selectToken } from '../models/accessTokenTable.js';
 
 dotenv.config();
 
@@ -59,4 +59,25 @@ export async function verifyRefresh(token) {
       message: err.message,
     };
   }
+}
+
+// refresh가 없을 때 refresh & access 토큰을 생성하고 쿠키에 저장
+export function setTokens(res, user) {
+  const accessToken = accessSign(user.id);
+  const refreshToken = refreshSign(user.id);
+
+  res.cookie('accessToken', accessToken, {
+    secure: true,
+    httpOnly: true,
+    expires: new Date(Date.now() + 3.6e6), // 1시간 뒤 만료
+    sameSite: 'lax',
+  });
+
+  res.cookie('refreshToken', refreshToken, {
+    secure: true,
+    httpOnly: true,
+    expires: new Date(Date.now() + 1.2096e9), // 2주 뒤 만료
+    sameSite: 'lax',
+  });
+  insertToken(user.id, refreshToken);
 }
