+++ routes/login.js
@@ -1,4 +1,5 @@
-import loginController from '../controllers/loginController.js';
+import consoleLogger from '../lib/consoleLogger.js';
+import { setTokens } from '../lib/jwtUtils.js';
 import { verifyLoginUser } from '../middlewares/verifyUser.js';
 
 export default function loginRoute(express, passport) {
@@ -10,13 +11,26 @@ export default function loginRoute(express, passport) {
   });
 
   // 42 OAuth 2.0
-  router.get('/42', passport.authenticate('42'));
-  router.get(
-    '/42/return',
-    passport.authenticate('42', { successRedirect: '/login/token', failureRedirect: '/login' })
-  );
+  router.get('/42', passport.authenticate('42', { session: false }));
 
-  // access 및 refresh token 생성 controller
-  router.get('/token', loginController);
+  // return으로 redirect와 동시에 종료
+  router.get('/42/return', (req, res) => {
+    passport.authenticate('42', { session: false }, async (err, user) => {
+      try {
+        if (err) {
+          consoleLogger.error('passport authenticate return : error :', err);
+          throw err;
+        }
+        if (!user) {
+          consoleLogger.error('passport authenticate return : error : No User');
+          throw new Error('No User');
+        }
+        setTokens(res, user);
+        res.redirect('/');
+      } catch (err) {
+        res.redirect('/login');
+      }
+    })(req, res);
+  });
   return router;
 }
